[REF] popups, forms: moved all the popup jsx files into a folder called popups within the components folder
Additionally, changed the name of the Form to forms to match the name
convention of the popups folder
