+++ fitness_app/src/components/popups/Popups.module.css
@@ -1,0 +1,9 @@
+/* Title for Forms*/
+.title-form {
+	margin-top: 10px;
+	margin-bottom: 10px;
+	font-size: 2.5rem;
+	font-weight: bold;
+	color: var(--primary-color);
+	text-align: center;
+}
