+++ fitness_app/src/index.js
@@ -3,9 +3,9 @@ import ReactDOM from "react-dom";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import AddSubSession from "./components/AddSubSession/AddSubSession";
-import AddExercise from "./components/AddExercise/AddExercise";
-import AddUserWeight from "./components/AddUserWeight/AddUserWeight";
+import AddSubSession from "./components/popups/AddSubSession/AddSubSession";
+import AddExercise from "./components/popups/AddExercise/AddExercise";
+import AddUserWeight from "./components/popups/AddUserWeight/AddUserWeight";
 import ExerciseRegistry from "./pages/Registry/ExerciseRegistry";
 import "./index.css";
 import PrivateRoute from "./utils/PrivateRoute";
